+++ CHANGELOG.md
@@ -5,11 +5,35 @@ All notable changes to the [Nucleus Python Client](https://github.com/scaleapi/n
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.14.25](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.14.25) - 2022-11-04
+## [0.14.27](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.14.27) - 2022-11-04
 
 ### Added
 - Support for scene-level external evaluation functions
 - Support for uploading custom scene-level metrics
+
+
+## [0.14.26](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.14.26) - 2022-11-01
+
+### Added
+- Support for fetching scene from a `DatasetItem.reference_id`
+Example:
+```python
+dataset = client.get_dataset("<dataset_id>")
+assert dataset.is_scene  # only works on scene datasets
+some_item = dataset.iloc(0)
+dataset.get_scene_from_item_ref_id(some_item['item'].reference_id) 
+```
+
+
+## [0.14.25](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.14.25) - 2022-10-20
+
+### Updated
+- Items of a slice can be retrieved by Slice property `.item`
+- The type of items returned from `.items` is based on the slice `type`:
+  - `slice.type == 'dataset_item'` => list of `DatasetItem` objects
+  - `slice.type == 'object'` => list of `Annotation`/`Prediction` objects
+  - `slice.type == 'scene'` => list of `Scene` objects
+
 
 ## [0.14.24](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.14.24) - 2022-10-19